# PathFinder Backend - RAG Documents API.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PathFinder Backend - RAG Documents API",
    "description": "OpenAPI spec for PathFinder Backend RAG document management. Import this file or the served URL into Apidog to test JSON and PDF upload RAG flows.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/chat/sessions": {
      "post": {
        "summary": "Create chat session",
        "deprecated": false,
        "description": "",
        "operationId": "createChatSession",
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI1MTlmNTBkNC1iMWE4LTQ3ZDktYmJmNi0wODhhM2ZiOGM4ZTMiLCJlbWFpbCI6InN0dWRlbnRAZXhhbXBsZS5jb20iLCJyb2xlIjoidXNlciIsImlhdCI6MTc4MTYzNTE3MCwiZXhwIjoxNzgxNzIxNTcwfQ.Lc1BMiLZvsagCOSqHhIlYnSCqbBidroEsjgmNfHKqwA",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatSessionRequest"
              },
              "example": {
                "title": "Career chat"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Chat session created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionResponse"
                },
                "example": {
                  "session": {
                    "id": "20000000-0000-0000-0000-000000000001",
                    "user_id": "10000000-0000-0000-0000-000000000001",
                    "title": "Career chat",
                    "status": "active",
                    "created_at": "2026-06-16T10:00:00.000Z",
                    "updated_at": "2026-06-16T10:00:00.000Z"
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/BadRequest",
            "description": "Invalid request"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError",
            "description": "Internal server error"
          }
        },
        "security": []
      },
      "get": {
        "summary": "List chat sessions",
        "deprecated": false,
        "description": "",
        "operationId": "listChatSessions",
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "",
            "required": true,
            "example": "20000000-0000-0000-0000-000000000001",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI1MTlmNTBkNC1iMWE4LTQ3ZDktYmJmNi0wODhhM2ZiOGM4ZTMiLCJlbWFpbCI6InN0dWRlbnRAZXhhbXBsZS5jb20iLCJyb2xlIjoidXNlciIsImlhdCI6MTc4MTYzNTE3MCwiZXhwIjoxNzgxNzIxNTcwfQ.Lc1BMiLZvsagCOSqHhIlYnSCqbBidroEsjgmNfHKqwA",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat sessions fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionsResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/BadRequest",
            "description": "Invalid request"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError",
            "description": "Internal server error"
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "CreateChatSessionRequest": {
        "type": "object",
        "required": [
          "userId"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "example": "Career chat"
          }
        }
      },
...
      "ChatSession": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "archived",
              "deleted"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ChatSessionResponse": {
        "type": "object",
        "properties": {
          "session": {
            "$ref": "#/components/schemas/ChatSession"
          }
        }
      },
      "ChatSessionsResponse": {
        "type": "object",
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatSession"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "Success": {
                "summary": "Success",
                "value": {
                  "error": "userId is required"
                }
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "Success": {
                "summary": "Success",
                "value": {
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the accessToken returned from /api/v1/auth/register or /api/v1/auth/login."
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": []
}